+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## next version
 
+### Changed
+- Support for `XCConfig` project-relative includes https://github.com/xcodeswift/xcproj/pull/238 by @briantkelley
+
 ## 4.1.0
 
 ### Added